pyproject.toml
[tool.poetry]
name = "SkipPay"
version = "1.0.0"
description = "Skip Pay demo project"
authors = ["dkoci <daevid226@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.3"
django-environ = "^0.9.0"
django-filter = ">=1.0.1"
django-guardian = ">=1.0.1"
django-postgres-extra = "^2.0.6"
django_rest_framework = "^0.1.0"
markdown = "^3.4.1"
gunicorn = "^20.1.0"
uvicorn = ">=0.20.0"
psycopg2-binary = "^2.9.5"
drf-yasg = {extras = ["validation"], version = "^1.21.4"}
httpx = "^0.23.1"
urljoin = "^1.0.0"
h2 = "^4.1.0"
html5lib = "^1.1"
bs4 = "^0.0.1"
fake-headers = "^1.0.2"
lxml = "^4.9.1"
django-bootstrap-v5 = "^1.0.11"

[tool.poetry.dev-dependencies]
Faker = ">=5.0.1"
pytest = ">=6.2.4,<7.0.0"
pytest-cov = ">=2.10.1"
pytest-django = "^4.5.2"
locust = ">=2.13.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
